Allow downloading PDFs instead of opening them in Brave.
Fixes brave/brave-browser#1531

The fix does the following:

1. Makes the initial decision on whether to load PDFJS extension based
on the value of kPluginsAlwaysOpenPdfExternally profile preference in
addition to the command line switch.

2. Watches profile preference value kPluginsAlwaysOpenPdfExternally and
adds/removes PDFJS component when the value changes.

3. Modifies js behind the chrome://settings/content/pdfDocuments so that
if the PDFJS extension is disabled from the command line the option to
download PDF files instead of opening them in Brave is set to ON and the
toggle is disabled.

4. Adds an alternative call for whitelisted extensions in
extensions/common/manifest_handlers/mime_types_handler.h that removes
Chrome's PDF Extension ID from whitelisted IDs and patches
chrome/browser/plugins/plugin_utils.cc to use the alternative call. This
is done because on Linux that extension is not removed from profile
enabled extensions (because there is no extensions install verification)
and it ends up being picked as a handler for PDFs. If our PDFJS extension
is disabled via command line, this causes a PDF to be neither displayed
nor downloaded.

Note, that this fix doesn't address being able to turn off opening PDFs
in Brave in Guest/Tor profiles. The web ui setting to do so is not
available in these profiles.

The command line switch to not load PDFJS already applies to all profile
types.

Adds browser tests to check that when the preference is set to always
download or PDFJS is disabled from command line a download is initiated
when navigating to a pdf url.